Refactoring and tests
